Start Dialog and abilities while started Dialog fixes.
Fixed awkard error in which player character while starting dialog at same time doing block animation. This was due no special handling for this kind of situation. Now it was fixed, so there is no stupid behavoiur like that.
Very similar situation was with attack from left mouse button and using abillities. Now it works as it should without this error.