Fixes for Game build
After making game build there were some issues to do.
First one was adding StreamingAssets folder, so the QuestDB in JSON format will be loaded in the game build.
Then there were needed some changes in DialogManager, because in the game build there were some methods that were not working properly.
In StartDialog there was issue with null objects of Player and Quest Giver, which then corespond with checking if distance for dialog wasn't exceeded.
Now it works correct.