Roll animation and movement fix
After last PlayerInput update the movement went into endless state. Player will not stop running until Player do something to stop it. It was silly mistake - movement input inside if statement for any key. Now it's above it, so works out this loop for abilities and othe interactions.
Fixed roll animation so now it will not transit to self, so there will be not breaking roll animation to do next roll animation if Player pressed the key for roll one time after another.